requirements.txt
absl-py==2.0.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.2.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
comtypes==1.2.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
executing==2.0.1
fastapi==0.115.2
filelock==3.13.1
firebase-admin==6.5.0
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.12.2
gast==0.4.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe @ file:///C:/Users/Vinh/Downloads/MarkupSafe-2.1.2-cp311-cp311-win_amd64.whl#sha256=e55e40ff0cc8cc5c07996915ad367fa47da6b3fc091fdadca7f5403239c5fec3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.9
motor==3.6.0
mpmath==1.3.0
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
openai==1.38.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
paho-mqtt==2.0.0
pandas==2.2.3
parso==0.8.3
pbkdf2==1.3
pefile==2023.2.7
pillow==10.2.0
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.1.0
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
pymongo==4.9.2
pyparsing==3.1.1
pypiwin32==223
PySocks==1.7.1
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.2
pywifi==1.1.12
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
scipy==1.14.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
stack-data==0.6.3
starlette==0.40.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tokenizers==0.19.1
torch==2.1.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.32.0
watchfiles==1.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
